fix: Streamlit Cloud 배포 오류 수정
🐛 버그 수정:
- PostgreSQL 타입 불일치 해결: stock_invested를 ::NUMERIC으로 명시적 형변환
- NameError 수정: account_info → account 변수명 수정 (app.py line 980, 982)

오류 메시지:
- "Returned type double precision does not match expected type numeric in column 7"
- "NameError: name 'account_info' is not defined"

파일 수정:
- complete_schema.sql: get_cash_transaction_summary() 함수 stock_invested 형변환
- app.py: 현금 내역 탭 account 변수 참조 수정

+++ complete_schema.sql
@@ -418,7 +418,7 @@ BEGIN
              FROM transactions
              WHERE account_id = p_account_id AND currency = p_currency),
             0
-        ) as stock_invested,
+        )::NUMERIC as stock_invested,
         calculate_cash_balance(p_account_id, p_currency) as current_cash_balance
     FROM accounts a
     WHERE a.id = p_account_id;
